dalton: fix error where the user's house (and its sections) would not import
this happened because they added a new "greenhouse coworking space" group. this
brilliant code saw that (more specifically, saw the word "house") and decided
that, logically, the greenhouse coworking space must be the user's house.

it's terrible code and we make it more terrible by just excluding the word
greenhouse from it