data/calendar/Event: in CalculateTimes, take the start timezone into account
this is because the concept of a 'day' depends on the timezone, due to things
like daylight savings time

we do the tzdata lookup in CalculateTimes, which means it can error now. it's
possible we want to move that lookup somewhere else? there just doesn't seem to
be a great place to put it where it'd apply to all events

related to MyHomeworkSpace/client#167