calendar: fix bug where recurring events at certain times would end one day earlier than they should
the issue was that we were parsing the until date in UTC, but the actual event
times were in the user's timezone (probably America/New_York).

so, if on the event's last iteration it was past midnight on the until date in
UTC, the event would not be shown for that day. fixed by loading the date as a
ISO8601 string into the RecurRule, and then parsing it only when we actually
want to use it and we know what timezone to use.